% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize a tibble/data frame column with one-way hashing}
\usage{
anonymize(df, col, char_length = 10)
}
\description{
Anonymize a tibble/data frame column with one-way hashing
}
