% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggprev}
\alias{ggprev}
\title{Preview ggplot file}
\usage{
ggprev(
  file_name = util_defaults$DEFAULT_FILE_NAME,
  width = util_defaults$DEFAULT_WIDTH,
  height = util_defaults$DEFAULT_HEIGHT,
  open = TRUE
)
}
\description{
Preview ggplot file
}
