% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggupd}
\alias{ggupd}
\title{Update already previewed ggplot file}
\usage{
ggupd(
  file_name = util_defaults$DEFAULT_FILE_NAME,
  width = util_defaults$DEFAULT_WIDTH,
  height = util_defaults$DEFAULT_HEIGHT
)
}
\description{
Update already previewed ggplot file
}
